Move RankedMenu test to ui-components Follow-up of #10373 refactoring
